I wrote a script for auto installing
The link is secured to a single commit so the file cannot be edited. If preferred, you could copy the script into the readme.md or into a seprate file in this repo- I don't care, nor do I really want credit. As long as you notify me it was moved (I would like to know if my script is actually being used by people!)

This script automates the full process, including gathering and installing all dependencies (including git, cmake, and the normal pepper), pulling the repo off of git, creating all needed directories, building the source, copying the file to the firefox addons, and disposing of the extra files it created and source it downloaded.

I just had found the installation process to be a pain, and I thought people might want it automated. It also makes it so that more lazy people will use this plugin.

Also, the lazy people part would be more active if the link was at the top, but it's up to you to move it up there.