get back to the pkg_check_modules
Apparentely the find_package call fails to found the PulseAudio in some systems like debian, and the if/else conditionals was failing until i put the ${PULSEAUDIO_FOUND} in the argument. 
So i saw 2 ways to solve this problem, one was to make  a .cmake file with some configuration that allows cmake to found the PulseAudio libraries, or get back to the pkg_check_modules. the simplest way that i saw was getting back to pkg_check_modules.

And the conditionals were not working until i add the "${PULSEAUDIO_FOUND}" argument to the conditionals, i'm not really sure why, because it supposed to work on either of those two ways. but doing this the PulseAudio recognition was solved for me.

it's not much, but i hope it helps a bit.